last

+++ app/Http/Controllers/TasksController.php
@@ -16,20 +16,13 @@ class TasksController extends Controller
      */
     public function index()
     {
-    $data = [];
-    if (\Auth::check()) {
-        $user = \Auth::user();
-        $tasks = Task::paginate(25);
-        
-        $data = [
-            'user' => $user,
-            'tasks' => $tasks,
-        ];
-    }        
-
-    return view('welcome',$data);
+      if (\Auth::check()) {
+      $user = \Auth::user();
+      $tasks = $user->Tasks()->paginate(25);
+      return view('tasks.index',['tasks'=> $tasks,]);
     }
-
+      return view('welcome');
+    }
 
     /**
      * Show the form for creating a new resource.
@@ -38,10 +31,10 @@ class TasksController extends Controller
      */
     public function create()
     {
-    $task = new Task;
+    $tasks = new Task;
 
     return view('tasks.create', [
-    'task' => $task,
+    'task' => $tasks,
     ]);
     }
 
@@ -55,12 +48,12 @@ class TasksController extends Controller
     {
     $this->validate($request, [
     'content' => 'required|max:191',
-    'status' => 'required|max:10',
     ]);
     
     $task = new Task;
+    $user = \Auth::user();
     $task->content =$request->content;
-    $task->status =$request->status;
+    $task->user_id =$user->id;
     $task->save();
     
     return redirect('/');
@@ -74,11 +67,16 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-    $task = Task::find($id);
-    
-    return view('tasks.show', [
-    'task' => $task,        
-        ]);
+        $task = Task::find($id);
+        if($task === NULL){
+            return redirect('/');
+        }
+            if (\Auth::id() === $task->user_id) {
+                return view('tasks.show', [
+                'task' => $task,        
+                ]);
+            }
+        return redirect('/');
     }
 
     /**
@@ -107,13 +105,11 @@ class TasksController extends Controller
     {
         $this->validate($request, [
             'content' => 'required|max:191',
-            'status' => 'required|max:10',
         ]);
         
     $task = Task::find($id);
 
     $task->content = $request->content;
-    $task->status = $request->status;
     $task->save();
     
     return redirect('/');
@@ -127,9 +123,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-    $task = Task::find($id);
-    $task->delete();
-
-    return redirect('/');
+        $task = Task::find($id);
+    
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/');
+        }
+            return back();
+        }
     }
-}
